Reorganize ButtonGroup component and stories.
