+++ src/ButtonGroup/ButtonGroup.stories.tsx
@@ -1,8 +1,7 @@
 import React from 'react'
 import {Story, Meta} from '@storybook/react'
-import ButtonGroup from '../ButtonGroup'
-import {Button, ButtonProps, IconButton} from '../Button'
-import {PlusIcon, DashIcon} from '@primer/octicons-react'
+import ButtonGroup from './ButtonGroup'
+import {Button, ButtonProps} from '../Button'
 
 export default {
   title: 'Components/ButtonGroup',
@@ -48,10 +47,3 @@ Playground.argTypes = {
     },
   },
 }
-
-export const IconButtons = () => (
-  <ButtonGroup>
-    <IconButton icon={PlusIcon} aria-label="Add" />
-    <IconButton icon={DashIcon} aria-label="Subtract" />
-  </ButtonGroup>
-)
