+++ src/ButtonGroup/index.ts
@@ -1,0 +1,1 @@
+export {default, ButtonGroupProps} from './ButtonGroup'
